requirements.txt
fastapi
uvicorn
sqlmodel
pyjwt
passlib
python-multipart
bcrypt
python-dotenv
pydantic
pydantic[email]